+++ src/Utils/Utils.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace FINDOLOGIC\FinSearch\Utils;
 
 use Composer\InstalledVersions;
+use Exception;
 use FINDOLOGIC\FinSearch\Definitions\Defaults;
 use FINDOLOGIC\FinSearch\Findologic\Resource\ServiceConfigResource;
 use FINDOLOGIC\FinSearch\Struct\Config;
@@ -161,21 +162,31 @@ class Utils
         // Composer 2 runtime API uses the `InstalledVersions::class` in favor of the
         // deprecated/removed `Versions::class`
         if (class_exists(InstalledVersions::class)) {
-            $platformIsInstalled = InstalledVersions::isInstalled('shopware/platform');
-            if ($platformIsInstalled) {
-                return InstalledVersions::getPrettyVersion('shopware/platform');
+            if (InstalledVersions::isInstalled('shopware/platform')) {
+                if (InstalledVersions::getPrettyVersion('shopware/platform')) {
+                    return InstalledVersions::getPrettyVersion('shopware/platform');
+                }
             }
 
-            return InstalledVersions::getPrettyVersion('shopware/core');
+            if (InstalledVersions::isInstalled('shopware/core')) {
+                if (InstalledVersions::getPrettyVersion('shopware/core')) {
+                    return InstalledVersions::getPrettyVersion('shopware/core');
+                }
+            }
         }
 
-        $packageVersions = Versions::VERSIONS;
-        $platformIsInstalled = isset($packageVersions['shopware/platform']);
-        if ($platformIsInstalled) {
-            return $packageVersions['shopware/platform'];
+        if (defined('PackageVersions\Versions::VERSIONS')) {
+            $packageVersions = Versions::VERSIONS;
+            if (isset($packageVersions['shopware/platform'])) {
+                return $packageVersions['shopware/platform'];
+            }
+
+            if (isset($packageVersions['shopware/core'])) {
+                return $packageVersions['shopware/core'];
+            }
         }
 
-        return $packageVersions['shopware/core'];
+        throw new Exception('Used Shopware version cannot be detected');
     }
 
     protected static function cleanVersionCommitHashAndReleaseInformation(string $version): string