+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [SW-569] The option to enable Findologic on category pages, has been moved to the API section.
 - [SW-678] Fixed a bug that caused console errors on Direct Integration pages, where the URL did not contain a query parameter.
 - [SW-624] Fixed a bug that would cause an error if the Shopware ElasticSearch bundle is not available.
+- [SW-681] Fixed a bug that caused the Shopware version not to be properly detected on some Shopware installations.
 
 # 2.6.0
 - [SW-673] Improved performance on category pages, by improving the way the plugin gets the current category.